I understand modules/packages a bit better now
