+++ src/tests/environment_variable_evaluation/test_environment_variables_braces.py
@@ -1,6 +1,6 @@
 import unittest
 import os
-from tests.compose_generator import ComposeGenerator
+from ..compose_generator import ComposeGenerator
 
 
 class TestBracesNoUnderscoreNoDigitVariableInterpolation(unittest.TestCase):
