+++ src/tests/service/test_service_credential_spec.py
@@ -1,6 +1,5 @@
 import unittest
-import os
-from tests.compose_generator import ComposeGenerator
+from ..compose_generator import ComposeGenerator
 
 
 class TestServiceCredentialSpec(unittest.TestCase):
