+++ src/tests/service/test_service_deploy.py
@@ -1,6 +1,5 @@
 import unittest
-from pycomposefile import ComposeFile
-from tests.compose_generator import ComposeGenerator
+from ..compose_generator import ComposeGenerator
 
 
 class TestComposeServiceDeploy(unittest.TestCase):
